v2/requirements.txt
# Bottle Checker V2 — Dépendances
opencv-python>=4.8
numpy>=1.24
ultralytics>=8.0
Pillow>=10.0
paho-mqtt>=2.0
flask>=3.0
